acho que arrumou o bug de travar

+++ headers/tempo.h
@@ -75,12 +75,13 @@ int deltaTempo(int timeOut,struct timespec start){
     clockid_t clk_id = CLOCK_MONOTONIC_RAW;
     clock_gettime(clk_id,&atual);
     if((atual.tv_sec-start.tv_sec)>0) {
-        deltaT =(atual.tv_nsec/1000)+1000000-start.tv_nsec/1000;
+        deltaT = (atual.tv_nsec/1000)+1000000-start.tv_nsec/1000;
     }
     else {
-        deltaT =(atual.tv_nsec/1000)-start.tv_nsec/1000;
+        deltaT = (atual.tv_nsec/1000)-start.tv_nsec/1000;
     }
-    if (deltaT >= timeOut*1000){
+    if (deltaT >= (long int)timeOut*1000){
+        //printf("Atual: %ld\n", atual.tv_nsec);
         retorno = 1;
     }
     return retorno;
